金額合計値とコードのリファクタリング
