+++ AzureCsvApp/AzureCsvApp/AzureCsvApp.csproj
@@ -53,14 +53,20 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="csv\CsvDataMap.cs" />
     <Compile Include="Form1.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="Form1.Designer.cs">
       <DependentUpon>Form1.cs</DependentUpon>
     </Compile>
+    <Compile Include="model\CsvData.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="purocess\LeadSumPrice.cs" />
+    <Compile Include="purocess\Inport.cs" />
+    <Compile Include="purocess\SumPrice.cs" />
+    <Compile Include="sql\Conection.cs" />
     <EmbeddedResource Include="Form1.resx">
       <DependentUpon>Form1.cs</DependentUpon>
     </EmbeddedResource>
